refactor(cli): extract creation-ledger command group to cli/ module (#1532 Phase B)
Second command-group extraction. Moves the 10 creation-memory ledger
commands (resume-bundle status/record/build, commitment
status/record/set-state/lifecycle, work-product
status/record/recall-search) from registerCli into
cli/creation-ledger-commands.ts behind a registerCreationLedgerCommands
registrar. Same behaviour-preserving, verbatim-move pattern as the
research-status extraction; Phase A contract tests stay green.

cli.ts: 9663 -> 9363 LOC (ratchet baseline updated; 9872 -> 9363 total).
